Moves removing invalid filter graphql schema into its own file
